ecnf base change part 2
