+++ lmfdb/ecnf/templates/ecnf-curve.html
@@ -503,12 +503,16 @@ Its isogeny class <a href={{ ec.urls.class }}>{{ec.short_class_label}}</a>   con
 <h2> {{KNOWL('ec.base_change','Base change')}} </h2>
 <div>
 <p>
-This curve {% if ec.base_change %} is the
-{{KNOWL('ec.base_change','base change')}} of
-{% for E0 in ec.base_change%}
-<a href={{url_for("ec.by_ec_label",label=E0)}}>{{E0}}</a>,
-{% endfor %}
-defined over \(\Q\), so it is also a {{KNOWL('ec.q_curve', '\(\Q\)-curve')}}.
+This curve {% if ec.base_change_Q %} is the
+{{KNOWL('ec.base_change','base change')}} from \(\Q\) of
+<ul>
+{% for E0 in ec.bcQtext%}
+  <li>
+{{E0 | safe}}
+</li>
+  {% endfor %}
+</ul>
+It is hence also a {{KNOWL('ec.q_curve', '\(\Q\)-curve')}}.
 {% else %}
 is not the {{KNOWL('ec.base_change','base change')}} of an elliptic curve defined over \(\Q\).
 {% if ec.qc=="yes" %}
@@ -520,6 +524,19 @@ It has not yet been determined whether or not it is a  {{KNOWL('ec.q_curve', '\(
 {% endif %}
 {% endif %}
 </p>
+{% if ec.base_change_NF %}
+<p>
+{% if ec.base_change_Q %} It is also {% else %} This curve is {% endif %} the
+{{KNOWL('ec.base_change','base change')}} of
+<ul>
+{% for E0 in ec.bcNFtext%}
+<li>
+  {{E0 | safe}}
+  </li>
+{% endfor %}
+</ul>
+</p>
+{% endif %}
 </div>
 
 {% if DEBUG %}
